--- src/main/java/com/kosa/controller/CartControllerTest.java
@@ -1,25 +1,0 @@
-package com.kosa.controller;
-
-import org.springframework.stereotype.Controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-
-
-@Controller
-public class CartControllerTest {
-    
-
-    @RequestMapping(value = "/cart", method = RequestMethod.GET)
-    public String home() {
-                
-        return "cart/cart";
-    }
-    @RequestMapping(value = "/cartt", method = RequestMethod.GET)
-    public String home2() {
-                
-        return "cart/test";
-    }
-    
-    
-}
